Initial commit
The start of a new adventure ...

This is the initial commit of Metaindex, an idea that has been running
through my head for several months now. Metaindex will be a library that
scans a directory of flac files, extracts the metadata, and writes a
database of metadata with indexes to support efficient lookup. The idea
is that this library could be the basis for a music player, but it will
be useful for different purposes as well. I intend to use it for Fmcount
for instance, where it would be useful just for determining track-album-
artist relationships.

My current idea is to have something with an on-disk format that can
be mmapped. That way the memory footprint would be low on memory-
constrained systems, and performance could be high when sufficient
memory is available. In particular, I would like to be able to use
this as a basis for a fast local music player. There are a few okay-ish
music players out there, but most choke on a music library of my size
(about 13k tracks currently). I also want to build a network music
player that can run on a Raspberry Pi, so memory usage is a concern.
Generating the index should ideally be possible on such a low-end
device, and consuming it on there is a must-have. And the index should
be portable as well, so I can just rsync it everywhere in case building
it on a Raspberry Pi is too slow.

So let's sum up the requirements:

 * Low memory footprint when reading the index.

 * Efficient lookup of track <-> album, album <-> artist, lookup by
   Musicbrainz id.

 * Efficient listing of all artists, of all albums by an artist, of all
   tracks in an album.

 * Efficient fuzzy search on track title, album title, and artist.

 * Portable across different machines by rsyncing.

 * Incremental build, no rescan of all files when one album has been
   added. A rebuild of some of the files in the index, after adding an
   album, is ok. But the music files themselves cannot be accessed.

 * Possible to build in a streaming manner under memory constraints.

The latter two points can be achieved by building small indices in
memory, flushing those to disk, and then having a merge operation that
can be streamed. This would support both the "append a single album" use
case, as well as "index a large library through divide and conquer".