Avoid crash when metadata is missing
Also, experiment results: all of the metadata for my library, as
produced by this utility, is 1.6 MiB in size. That is practically
nothing, which is great, because it means all metadata fits easily in
memory, and even doing a linear scan to search would be feasible.
(Though I am not going to do that, of course.) There would be a extra
data required to store album metadata, but on the other hand, a lot of
bytes can be saved by storing the uuids in 16 bytes rather than 36, and
not all of this info needs to be stored for every track.

But this does open op new possibilities: it means flushing to disk and
merging is probably not necessary, and building the index would be
possible even on low-end devices while keeping everything in memory.