Add a few struct definitions
I've been thinking about what data structures to use for days. I want
something which is easily mergeable, and which allows efficient lookups.
Initially I did not want to index by UUID (which might require a search
rather than O(1) lookup), using offsets instead. But I could not find a
way to make that fit without seriously complicating merging indices
(which is required for efficient updates and construction). In the end,
the approach of a traditional database does not seem so bad. What I will
build is basically a database with hard-coded schema. There would be
sorted indices as well. It can be merged efficiently, and it can be made
reasonably small.