Implement poor man's non-blocking IO
This speeds up indexing in my case by another 25% or so. Down from the
original ~130 seconds to ~80 seconds, just by changing the number of
threads that do IO, and changing the order in which they do it.

It might be even better to just protect the path iterator with a mutex,
and have the multiple threads "pull" (or "steal") from there, rather
than busy-waiting from the main loop to put the path in some inbox. If
disk IO is the bottleneck, the mutex would not be contended. For the
cached case it might be when the number of threads is high. More
measurements are needed.