Make threads pull from mutex, not channel
It is simpler, and also the most performant solution I've found up to
now. Furthermore, when all data is in the disk cache, it degrades less
than the previous approach.

The downside is that I have to pull in another dependency for this to
work, as std dropped the scoped thread pool. Fortunately Crossbeam does
not pull in any third-party dependencies; this one is manageable.