Remove "sort artist" field
It is not useful for what I want to do; you can browse by album artist,
and search by track artist, but sorting all track artists is not
something that I plan to support.

This also paves the way for making two Track structs fit in one cache
line, if I save some bytes on disc and track number, and duration.