Measure string duplication
It turns out that about 22% of the strings (title, artist, album artist,
albu artist sort, and album title) need to be inserted. In the other
cases, they existed already, so there is a lot of duplication. This is
not surprising: album, and album artist information is repeated for
every track in the album, an artist information even across every track
of the discography. But it is always better to measure than to guess. So
now I have the data.