Do allow duplicate tracks with the same id
Sometimes a track is released on multiple albums, but it might be the
same track. Musicbrainz assigns it the same id in that case. But I want
different entries in the index, because they are different files, and
the track numbers differ.

Now there is no very strong reason why the track id should be unique.
Because the album id is right after it in memory, it is possible to use
the (track id, album id) as a unique id. The downside is that it is a
bit more expensive to check, and having non-unique ids is a bit tedious.

One alternative would be to xor the track id with the album id, so the
same track has different ids on different albums. The downside of that,
is that it is no longer possible to look up a track if you only have the
Musicbrainz id. But on the other hand, that is kind of the point: the
track id alone is not unique on Musicbrainz. Actually, as I am writing
this ... perhaps xor is better after all. If required we could have an
extra table to track the mapping.