Base track id on album id only, not artist
52 bits is enough to identify 50k albums uniquely, with a collision
probability of 0.00003%. We could actually abuse this to make all album
ids 52 bits (with zeros at some point), and that would allow collapsing
the disc + track and parent album info all into one, saving 10 bytes
from the 32-byte track struct. That is not enough to fit three of them
into a cache line though, so it is probably not worth it, because then
structs can straddle a cache line again. Although still more of them
would fit in the cache, so perhaps it would be worth it after all.