Use Hyper not Simple Server
Simple Server crashed when I fed it a url that contained "%20". Okay,
I'll switch to something else then. The Hyper API looks simpler than the
Iron one, so let's go with Hyper. It has the additional benefit that it
can handle multiple connections simultaneously, although I do not think
that will be important for my use case.

Bring in the megabytes of dependencies supporting obscure vaporware
operating systems! But yeah, there does not appear to be a lightweight
server that actually works, so apologies, but it seems that Hyper is the
path of least resistance.