Define temporary MemoryMetaindex fields
I will need to merge the builders into something that can be used for
serving. Ideally that should be an efficient in-memory format that can
be written to disk, and later mmapped, but for the initial version, it
is fine to use bulkier data structures and rebuild them at program
start.