Make text output location configurable
Writing to stdout should not be hard-coded, it spams the console when I
try to benchmark something. There is still more printing to stdout
hardcoded that I should get rid of.

In the end though, the channel and structured issues should be
sufficient. The end-user program could print those to stdout, or it
could ignore them.