Optimize artist lookups with bookmarks
The idea is to keep a small table, to reduce the range that we need to
binary search, hence cutting down on the number of cache misses.

Lookup time went down from 13ns to 4ns, so this is quite a success.
On my Skylake i7 that is at least, it remains to be seen how this
translates to low-end ARM cpus.