Implement loading album tracklist
Navigation does not work, and this is unstyled, but other than that, it
works! What is even more surprising, with Elm, once it compiled it did
really work correctly on the first time.

As a drive-by change, make ids a bit more type safe in Elm by
distinguishing between album ids and track ids.