De-emphasize parenthesized parts of track titles
This is just a heuristic, but I think it is quite nice. Do not set the
parenthesized part of a track title in bold. It goes wrong for example
on Oasis' "It's Getting Better (Man!!)", but for "All Around the World
(reprise)" on the same album, it works perfectly. So I guess this is
fine for now.