Update readme, rename project to Mindec
There are already a number of things called "Metaindex", and the name is
indeed quite general. After a few weeks of thinking about this, the best
I could come up with was "Mindec". I don't think there is a project with
this name already, and it is still a bit reminiscent of "music metadata
index".

Let's rename all of the other things later. I might still change the
name if I need to, for some reason I am not completely sold on Mindec.
Naming things is hard.