Implement image thumbnailer
It is based on an external process for isolation: I could use a Rust
library, but the available ones are not that mature, and one crash would
take down the entire process; one malicious input would compromise the
server (although admittedly that is not my threat model -- input files
are trusted). Imagemagick is mature and widely available.