Serve thumbnails in library browser
This makes the page load pretty much instantly -- a surprise to me,
given that there are about 1000 distinct images on there, and I thought
Hyper does not implement Connection: keep-alive. I might be wrong about
the last part. Still, it is good to know that there is room for
improvement (also, sendfile would be a win, I hope), and yet the current
version is fast enough.