Strip metadata from cover art thumbnails
It turns out that many of the cover art images embedded in my flac files
contain metadata. Most of it is a generic date time field, occasionally
camera metadata. Sometimes it is a professional camera, for covers based
on a photograph. Not sure if this is intentional or an oversight. In
other cases the camera is an iPhone, where I apparently took somebodies
picture of a cover as the cover art image. In yet other cases, the cover
art image contains a note that is actually related to the cover art.

Anyhow, this data is not useful when displaying thumbnails, but it
does cost bytes. The library browser can load faster if we remove
this metadata.