Pin Rust toolchain version
I don't want to be bitten by all of these breaking changes again. Pin
all the things. Perhaps I should use Nix, but using Rustup is a bit more
flexible for development.

I picked 1.25.0 because that is the version that Ubuntu 18.04 LTS ships,
that will be a good common denominator in the future, without being
sereverly outdated already at this point.