Do not overwrite conflicting metadata
Previously the new metadata would be stored unconditionally. If for
example you had two albums with the same album artist mbid but different
names credited, the name in the builder would flip between the two,
which makes the errors flip too. Also, the order of the names in the
error was the wrong way around, which is now fixed.