Add Purescript hello world
After a trip trying to build an Android app (in order to access
Android's Cast API), I discovered that Android app development is a huge
mess of multiple layers of binaries that download other binaries, that
are not available as official packages, only by downloading from the
website. Then you need piles of boilerplate, just to get the app going.

I thought I'd try Flutter, but it does not solve anything; you have all
of the trouble of regular Android development, and on top of that you
have all of the trouble of Flutter: it manages its own toolchain of
binaries from who-knows-where, it demands a writable installation
location, it introduces yet another build system and the Dart package
manager, and it comes with its own set of boilerplate.

Good luck trying to get something like this to build two years from now.
Good luck trying to get it to build on a different machine at all,
without first having to manually tune a lot of config.

So after that detour, I figured that perhaps a web-based app was not
such a bad idea as it seemed at first sight. Sure, the Cast API in
Chromium does not work on Arch, but I should be able to get it to work
on my phone at least. Then I don't need piles of unreproducible ad-hoc
tooling, just a single html page with a bit of Javascript.

Now, to add a complication: Javascript is a terrible language and I
would not want to burden this world by creating more of it, so we need a
sane language that compiles to JS (because wasm is still not useful by
itself). But that immediately introduces many problems again: the
compiler needs to be easy to install. The package manager, if any, must
be sane and lock dependencies. JS in the browser is bad enough, JS
outside of it is madness, so everything based on Node is out. That
leaves a few options:

 * Dart can compile to JS. On its own, without all of the Flutter stuff,
   it looks somewhat reasonable.

 * Elm compiles to JS. I was using it before, it works well, it doesn't
   have any weird dependencies. The tooling is great, but the language
   is a bit limited, and especially interfacing with the Cast API is
   going to be a pain.

 * Purescript compiles to JS. The tooling looks reasonable and is
   available in the AUR. Last time I checked, you needed Node-based
   package managers, but now there is psc-package which works well, and
   which freezes all dependencies, Stackage style. With the help of a
   friend to fill in the gaps in the documentation, I now have a
   makefile that builds a hello-world JS file, without tons of
   boilerplate. Let's see how far we get with this one.