Send cast request with metadata
Unfortunately, this only works from localhost, as Chrome now requires
https to make the Cast API available (localhost is a special case).

This is going to be problematic, as obtaining a certificate for a local
server is going to be a pain (although it might be possible with acme
dns challenges). A self-signed certificate is even more painful, because
I have to get it installed.

Alternatively, I could host the player file on a static page online, but
point it at the API server in my local network.