Explain external contributions are not accepted
Mindec is a personal project of mine, provided as-is, source-available,
and I have no intention at this point to adapt it to other people's
needs, or to maintain features that I have no use for myself.

Mindec is free software: users have the freedom to run it as they wish,
for any purpose. They have the freedom to study how the program works,
and to change it so it does their computing as they wish. They have the
freedom to redistribute copies so they can help others, and they have
the freedom to distribute copies of their modified versions to others.

However, Mindec being free software does not mean that users are
entitled to support, or that I feel obliged to integrate and maintain
changes that others have made.

I welcome changes that benefit me personally, but it is not immediately
clear for others what those might be. Especially given that Mindec is
opinionated in how it expects files to be tagged, and given that it has
a user interface which is subject to personal preference, I am afraid
that bikeshedding will ensue. One way to deal with the fact that
different people have different preferences, is to make the points where
preferences differ configurable and extensible. Many music players are
extremely versatile, and I have been a happy user of such software for
many years. But flexibility comes at an implementation and maintenance
cost, and even if an external contributor would take on the initial
implementation, over time the maintenance burden dominates. At this
point Mindec solves a personal need for me, but I do not have the time
to make it solve everybody's needs.

To avoid disappointing potential contributors, I think it is a good idea
to communicate expectations up front. And to avoid ambiguities, it is
better to err on the side of not giving potential contributors false
hope of their contribution being included. Therefore, for now the
contribution guide says that no external contributions apart from bug
reports are included. Even bug reports are a slippery slope, as the
distinction between bugs and features is not always clear, and I do not
care so much about bugs as long as they do not affect me personally. But
on the other hand, I do not want to reject help by well-intentioned
users either, so I will leave the issue tracker open.