Set up Travis CI
Yeah, I know, Travis is not really fashionable any more. But boring is
fine, at least I understand Travis pretty well. It is not fast, but it
is not mission-critical for me. They regularly break things and
their build environment is outdated, but I depend on Nix to manage
dependencies, which mitigates most breakage, and makes me independent
of their release cycle. An if necessary, Nix makes it easy to move to a
different CI provider. For now I just want to get a build status badge.