Take pinned Claxon from Github
I pointed it at a local path because it is easy for development. But I
am probably not going to revisit the Claxon API soon. It is suboptimal,
but my first priority now is to get Mindec in a working state, and we
can improve Claxon later. To get things to build on CI, I can't have
dependencies on local files, and for reproducibility it is essential to
not have those anyway. So take Claxon from Github.