Delete unused Chromecast code
At some point the idea was to have the server itself interact with the
Chromecast, to make it a kind of proxy, where it exposes an http API
on top of the somewhat convoluted Chromecast protocol.

But it turns out that the way in which the library I was using worked,
made it difficult to apply it to a real application: it could either do
a blocking wait for events, or it could send a command. But if it waits
for updates, it can't send a command. Which makes it difficult to e.g.
pause a track while it's playing. Also, the mdns discovery had some
issues.

Now I am building a web app that relies on Chrome's built-in Chromecast
support (after a brief detour with the idea of doing the same but with
an Android app). So far this is the first thing that actually works
reliably. I think this is the best way going forward. Perhaps in the
future we can revisit this, because the Chrome-based API is not ideal
either. But for now, no casting from Rust.