Add sort of scrollbar
Ugh, this is so cumbersome in html/css/js. This is one of the things
that would have been trivial in an app, yet is very painful on the web.
Halogen doesn't help here though, perhaps it is the wrong model for this
kind of app.

One solution I settled for now that is not great but somewhat usable
without too many hacks is to just have a button for each year, and they
stack along the right side, so you can tap anywhere to go there. Only on
mobile hover does not work as nicely. And really, on desktop too, it
would be nice to have a single balloon following, rather than many
balloons with hover effects. But it is something at least.