Add monad for building html
A reader monad for the parent element seems like quite a nice way to
build an html dsl. Probably the Halogen authors explored this before,
and there is a good reason why this does not work so well, and why it
uses lists for child elements ... but so far, this works, so let's go
with it and see where we end up.