Wrap album list construction in a tree
My (unscientific, based on nothing but thin air) hypothesis is that
Chrome is sometimes slow to handle my page, because it contains a huge
list of ~thousand elements. Certainly animating max-height at the top of
the page is less fluent than doing so at the bottom of the page,
probably because it causes layout of everything below. So then by
building a tree, such modifications can remain more local, as nodes that
haven't changed retain their layout with respect to their parent.

At least, that is my theory. So far I haven't felt any difference, and I
haven't measured properly either. So basically, this is a premature
optimization.

Also, it breaks styling as it makes the <li> not be direct children of
the <ul>.