Add function to get the queue
Only it does not work, because the Chrome Sender API is broken. :( It
limits the number of items in the queue it reports to 3, even though the
cast device internally may have many more items in the queue.

Also, the queueData field, which is documented, is always undefined
as far as I can tell.

So far this Chromecast adventure. Perhaps I need to consider an Android
app in Java after all.