Set up Android development environment
So it turns out the Chrome sender API is broken. And if felt pretty
second-class anyway (for example, it prints a deprecation warning, for
which a bug exists that has not been replied to for months, and the
initial removal date for the feature has already passed). Perhaps I
should give Android a second chance. It turns out we *can* manage the
Android SDK with Nix, so that brings some sanity. If I can also set up
a simple app with less than 10 boilerplate files, I will be happy.