Abort on panic
I just wasted half an hour debugging a deadlock, which turned out to be
caused by a panic (and therefore the main thread waiting for progress
that would never happen), but the message was buried between hundreds of
lines of output.

If a single thread cannot recover from the panic, why should the program
be able to recover from one of its threads panicking? It cannot; abort
the program on panic.