Add entry point for building Noblit database
I want to toy around with Noblit a bit, to see it scales, and also to
inform the design of the Rust API. One nice thing to try would be to
build a Noblit database out of all my music metadata. So let's add an
entry point for doing that, and a dependency on Noblit.