Bump transitive dependencies
I just deleted the Cargo.lock and did a fresh build. No idea what I am
getting now, and even if every package would provide proper changelogs,
no way I am going to read all of them for so many dependencies. This is
just dependency russian roulette. But the ecosystem does not leave me
much of a choice. Not updating is a possibility of course, but the
versions I pinned initially are just as arbitrary as these.