Implement basic search functionality
It only searches on entire words, not prefixes, and the data structure
for search is terrible. Also, it unions (without even deduplicating)
results for multiple words, instead of intersecting them. And it does
not handle url parsing very well.

But aside from that, this povides the foundation for search! And because
I already put a lot of effort into normalizing words, it actually works
pretty well!