Handle percent decoding of search query param
This is such a mess and so much more complicated than it needs to be.
When I replace the server with my own hand-rolled one, I should probably
just take the search query in the body, so I don't have to deal with
this percent encoding mess, which is especially complicated by Rust's
http ecosystem. It's utf-8 bytes in, it should be utf-8 bytes out. No
need to do expensive decoding involving *allocations*. But for now this
works. Now I can search for queries like "don't", with apostrophe.