Use css grids instead of floats
So, when you type in the search bar and then clear the text again, upon
deleting the last character, the entire album list reappears. And this
is slow, it causes a noticeable delay. Profiling reveals that the delay
is about 200ms, and that about half of it is recalculating style, and
the other half is doing layout.

Some Chrome help pages say that for apps to be faster, we have to reduce
the amount of work that happens in calculating style and doing layout.
And it claims that grids (or rather, flexbox) are faster to evaluate
than older techniques, which I was using. So let's roll with that.

So far, no obvious improvement, although I haven't done any scientific
measurements.

Also, this breaks the style of the search results, for now.