Print more stats when indexing is complete
For a mapping of artist id to list of album ids, I am wondering what the
best data structure is. And the answer is, it depends. I have many
artists in my library with only a single album. (Or alt least, where
there is only a single album by that artist in my library.) In that
case, an array of (artist id, album id) pairs makes a lot of sense. But
if there are many albums per artist, it may make more sense to have a
secondary array of album ids, and to have an array of (artist id, start
index, length) triples.

Currently, in my library I have 413 album artists and 1211 albums, so 3
albums per artist on average, although I suspect there are a few
outliers that pull up the average, and the median would be lower than 3.
Still, the order of magnitude should be right, and for the very prolific
artists, they may have 20 albums, but not 200, apart maybe from "Various
Artists" if you have many compilation albums.