Add release years to album list
I struggled for a while with where to place these. I'm still not
entirely satisfied, but this is the best I could come up with for now.
I do like to place albums in their historical context, so I do want to
have the release year in there. Putting it in front looks weird, because
it is not the most important thing in there. Aligning right does not
work on wide layouts. Putting it after the artist works reasonably well
because the artist is often shorter than the album title. And by making
it lighter, it is de-emphasized, and it is clearly not part of the
artist's name.