Don't open source file if thumbnail already exists
I noticed the "mindec cache" command being very slow after having
already indexed all music. Turns out, I was opening a file from every
album, and reading the cover already, and only then exiting if the
thumbnail already existed. But that means we still need to open the file
and read the entire cover from it, for *every album*, even if that album
already has a thumbnail. This is of course silly, we can stop
immediately if the target file exists.

Now running "mindec cache", with a warm page cache for the index itself,
is instant, if there are no new thumbnails to generate.