Remove alpha channel when generating thumbnails
I noticed after switching to Guetzli that for a few images, the color
was all of, as if it took the LAB values and interpreted them as RGB, or
as if RGBA was rotated into ARGB and then discarding the B channel, or
something like that. I'm not entirely sure what was going on, but one
problematic image contained an alpha channel. Previously Imagemagick
would have handled it upon conversion to jpg, but now that it exports as
png, the alpha channel can be preserved all the way, which is not what
we want. Getting rid of it all the way at the start resolves this.