Use new word index to power (prefix) search
That was quite simple in the end. I already have the BTreeSet, which is
the right format to turn into the word index later, so that's a simple
replacement in the constructor. And for search itself, that actually
becomes simpler too, because we no longer need to construct bounds or
unpack tuples. And by virtue of the prefix search of the word index, we
can now search for arbitrary prefixes, not just exact matches.