Write down thoughts on search function
I want to figure out what words I need to put in each index to build the
thing I want. I think that to support "each word occurs either in the
track title, or in the artist", I will need a mixed title-artist index.
Having two and combining them would lead to an exponential problem,
where there are 2^n ways of dividing the words over title and artist. I
don't know how to make that work with multiple indexes, apart from
enumerating all 2^n combinations, per combination do an intersection of
the results, and then take the union over combinations. But that gets
out of hand quickly. If they are all in the same index, then this is
accounted for. But to figure out which indexes I need, it helps to write
things down, which I am going to do next, but I first wrote down some
other search requirements.