Begin working on search module
The first step is a union iterator, that can yield the elements of
multiple slices (different words that share the same prefix, the search
query) in order. This is a basic step towards that.

I tried to work with slice iterators at first, but the types get
convoluted quickly, and they are not easier to use than indexing anyway,
because I need peek and zip. So just store indexes instead, I'll take
the boudns check hit. It can be removed later if needed.