Tune search ranking further
This is difficult, there are so many conflicting requirements. Maybe
I should collect a bunch of examples, rank them, and then fit a
polynomial that would best agree with my manual ranking. I suspect my
intuition also depends on the word itself. E.g. if "Time" occurs late in
the title, it can still be relevant, but if "Be" occurs late in the
title, it is not as specific, so it matters less. Perhaps word frequency
could weigh in as a proxy here.

For now, this will have to do.