Add dependency on purescript-aff-bus
I want to get the model/view under control a bit better, right now it is
essentially a mess of imperative code and callback hell (though Aff
makes callbacks very nice). I want more of a message loop architecture.
The simplest case of course would be a message loop acting on a state,
and a render function that turns the state into the page, a la Elm.
However, this breaks as soon as you care about the identity of your DOM
nodes, for example, because they have a css transition on them, so
deleting and creating a new one is not the same as recycling one. So
what I want to try instead is to still have a simple data model with
event loop that updates it, but to have a state alongside it with things
like DOM nodes, and a function that tries to make that state look more
like what the pure state says. I guess it is vaguely similar to the
virtual DOM and diffing idea, except I don't want it as an optimization,
I really want to implement the transitions by hand. Aff-bus seems like a
good way to tie back events that happen at the DOM level into events
that happen on a single event loop at the top level of the main program.