Add machinery for "recycler view" album list
This idea is inspired by [1], which targets infinite scroll. The album
list is finite, but the idea is the same: perform lay-out manually,
ensure that the DOM consists of only the visible nodes, and recycle
nodes and place them elsewhere when scrolling.

This adds much of the machinery to be able to do that, but not yet any
scroll handlers. Also, we don't store the scroll state in the app state,
which would be required to recover it on navigation.

[1]: https://developers.google.com/web/updates/2016/07/infinite-scroller