Make album list update more efficient
Previously, if the range changes quickly, we would have to incrementally
rotate until we catch up. Now we only keep the interection of the target
and current range, and recycle anything that falls outside. If the
target moves quickly, that means we recycle everything.

It turns out that performance does not have to come at the cost of
complexity, this new solution is less lines of code, and distinguishes
less cases, yay!