Split out setting <img> properties
setAttribute showed up as hot while profiling, but now that I split this
out, this shows up as hot. It is really assigning the src, which in turn
triggers an image load, that is taking so much time, not really
PureScript overhead. Although small bits can make a difference, of
course.

I could try to recycle <img> elements, rather than deleting them and
creating them anew.

What also stood out in the profiler is that sometimes hangs during
scroll are due to a major GC. Recycling the <img> elements instead of
destroying them and adding new ones might make a difference too.

I might even opt for keeping a lazily initialized (and possibly eagerly
filled in the background) array of images in memory.