Experiment with inline play/enqueue buttons
This is meh, on mobile it eats up too much space. I'll go with a menu,
where you can choose enqueue or play next. Or maybe I will just go with
enqueue by default, and make an option for "play next instead" in a
snackbar-like popup. Or maybe I just need a "move to front" in the
queue, then a normal click would be enqueue, and that's it.