Add dependncy on "alsa"
I was wondering whether to use the "alsa" crate or just "cpal", but cpal
has a slightly larger footprint due to Windows support. It is also a bit
higher level. Probably I should use it, but I am going to try and see
how far I get with raw Alsa.

What about PulseAudio? In combination with Gnome it just works, usually,
but on a headless server that I want to use for my music player, I've
had a lot of trouble configuring it so the user running my deamon is
able to play audio. I don't need any of the "features", and I'm afraid
it might silently do unexpected things (such as resampling audio when
the sound card does not support the sameple rate, rather than notifying
me with an error), so I am going to stay away from PulseAudio.

Audio is going to be used for local (on the server) playback. The
Chromecast did not work out, it is far to unreliable and disconnects at
random times. So now I'll play back from an external sound card instead.