Generate image thumbnails in parallel
I wanted to generate my thumbnails anew on a different disk, but it was
quite slow, so I thought I'd parallelize this, which I wanted to do for
a long time anyway.

By the time I finished writing the code, sequential thumnail generateion
was also pretty much done. But hey, now it's parallel.