Improve CPU utilization of thumbnailing
Instead of waiting for the oldest process, we can first check if any of
the others is done. If they are all running, *then* we can start a new
process.

Even with 32 jobs though, this does not saturate my 8-thread machine.
I'm not sure why not. It might be bottlenecked on IO to read the images,
although if that were the case, I would not expect it to hit round 2 as
often as it does. Hmmm ...