Resize in a *really* linear color space
Turns out I was doing this wrong all the time ... LAB is not linear in
ImageMagick. XYZ and RGB are. I noticed because I now also use the
thumbnail as a backdrop when the full-size image is loading, and I
noticed a popping effect, especially in images with a dark background
and sharp bright lines or text, it would pop in much brighter when it
loaded.

This is a sign that resizing is done in a nonlinear color space: if the
value at 50% represents only 25% lightness, then in an area with 2
bright and 2 dark pixels, you would expect to downsample to 50%
lightness, but we get 50% of the value, which is only 25% lightness. So
thin bright lines on dark backgrounds get thinner, and dark text or
lines on a light background gets fatter.

I played around a bit with covers where this was very obvious, mostly
covers that fake a gradient with lines of different widths, or a similar
kind of pattern. (For example, Jazz by Queen.) XYZ is definitely better,
but sometimes too light. sRGB with custom gamma sometmies did a bit
better, but it really depends on the image. I found a gamma of 2.05 to
be too bright in some parts of an image, but too dark in another. The
curve can be tuned further, but I'm not going to spend so much time
doing that here.

What may also play a role, is that Chrome's upsampling is very likely
not done in a linear color space either, so the upsampling again darkens
images.

A downside of this is that sharp lines become more pixelated, not as
smooth as they used to be. I tried Lanczos2 instead of Cosine, but did
not see an obvious improvement there.