Use ACCESS_RW_INTERLEAVED for playback, not mmap
The direct mmap mode is not supported on my Raspberry Pi. There is also
a non-direct mode that supports mmap for writing, but it is not clear to
me what the advantage is. I don't know what alsa-lib does after the
writei call, but I still have my buffer in memory that I need to copy to
Alsa, so whether I copy my bytes into an mmap provided by Alsa, or
whether Alsa copies the bytes from a buffer that I had anyway, it
doesn't seem like there should be a difference between the two, although
the documentation is sparse, so I don't know if my guess is right.