Assume that all music is stereo
This assumption currently holds for my library, and it would simplify a
lot of things if I can make this assumption (I don't have to dispatch on
the number of channels everywhere, and think about downmixing or filling
all channels, etc.). Given that I apparently have no non-stereo files, I
feel safe hard-coding this assumption. If an old album is mono, it can
still be encoded as stereo in flac. The mid-side encoding will be very
efficient because the side channel is 0, so not much is lost there. If
there is a fancy surround recording or anything like that ... well, I
don't have surround speakers, so I just would not get that version, and
add the normal stereo version to my collection instead.