Support switching sample formats in playback
I thought it would be easy, I could just use i32 all over the place, and
Alsa would take care of it ... but no luck, the i32 writing is a
convenience offered by the "alsa" crate, and it is only supported for
the s32 sample type. So now I'll have to do my own byte writing, because
I do need to support s24.

The upside is saving a lot of memory, half of it in the common case of
16 bits per sample.