Fix decoding flow
Early on, we don't want to decode a large batch in one go, we want to
decode a small batch to ensure low latency. So take that into account.

Aside from that, using an Option and the size of the blocks for a queued
track to infer its decoding state is fragile. Instead, make a proper
state machine for it.

It works now! I can run the "play" command and it plays the single
queued file.